config: more specific ingress adapter definition
